+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed SearchTextBox to ChapterSearchTextBox
 - Renamed SpacingStackPanel to ChapterStackPanel
 - Renamed SplitButton to ChapterSplitButton
+- Renamed TimeBox to ChapterTimeBox
 ### Supported .Net Versions
 - .Net Core 3.0
 - .Net Framework 4.5